# CUDA: Accelerate MXFP4 table lookup using `__byte_perm`