This PR accelerates MXFP4 vector dot product calculations by optimizing the `get_int_from_table_16` function.

The previous table lookup logic is replaced with the more efficient `__byte_perm` CUDA intrinsic. 

This results in a significant performance improvement for gpt-oss using the MXFP4 format on NVIDIA GPUs.

**Benchmarks (RTX 3080 Ti, MXFP4 gpt-oss-20b):**


**Master Branch:**
```
CUDA_VISIBLE_DEVICES=1 ./build_origin/bin/llama-bench -m ~/workspace/models/gpt-oss-20b/gpt-oss-20b-mxfp4.gguf -fa 1 
ggml_cuda_init: GGML_CUDA_FORCE_MMQ:    no
ggml_cuda_init: GGML_CUDA_FORCE_CUBLAS: no
ggml_cuda_init: found 1 CUDA devices:
  Device 0: NVIDIA GeForce RTX 3080 Ti, compute capability 8.6, VMM: yes
| model                          |       size |     params | backend    | ngl | fa |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | -: | --------------: | -------------------: |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |  1 |           pp512 |      3719.22 ± 51.68 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |  1 |           tg128 |        153.88 ± 0.26 |

build: a094f3814 (6210)
```

**This PR:**
```
CUDA_VISIBLE_DEVICES=1 ./build/bin/llama-bench -m ~/workspace/models/gpt-oss-20b/gpt-oss-20b-mxfp4.gguf -fa 1 
ggml_cuda_init: GGML_CUDA_FORCE_MMQ:    no
ggml_cuda_init: GGML_CUDA_FORCE_CUBLAS: no
ggml_cuda_init: found 1 CUDA devices:
  Device 0: NVIDIA GeForce RTX 3080 Ti, compute capability 8.6, VMM: yes
| model                          |       size |     params | backend    | ngl | fa |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | -: | --------------: | -------------------: |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |  1 |           pp512 |      4127.83 ± 29.88 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |  1 |           tg128 |        176.38 ± 0.12 |

build: bd4b6cd23 (6211)
```