@IMbackK 2025-08-20
While HIP also supports `__byte_perm` performance impact is negative.