@lovedheart 2025-08-20
@IMbackK
ChatGPT's answer:

1. What __byte_perm does

On CUDA, __byte_perm(x, y, s) maps directly to the PRMT (byte permute) instruction on NVIDIA GPUs.

That instruction can pick arbitrary bytes from two 32-bit registers and shuffle them into a single 32-bit result — in one cycle on modern NVIDIA SMs.

So on NVIDIA, this is a **native** instruction.

2. On AMD / HIP

HIP provides __byte_perm as part of its CUDA compatibility, but there’s no direct equivalent PRMT instruction on **GCN/RDNA ISAs**.

The compiler has to **emulate** the operation: usually with a sequence of bitfield extract (BFE), shifts, masks, and ORs.

That expands into **multiple instructions (5–12+ depending on pattern)**.

**Throughput and latency are much worse than the single-cycle CUDA version.**