@IMbackK 2025-08-20
actually while `__byte_perm` is indeed implemented as bitshifts in hip, AMDGCN has a similar instruction that could be used: V_PERM_B32, the compiler is already using V_PERM_B32 and a quick implementation that uses V_PERM_B32 by hand like this pr dose for nvidia is worse than what the compiler is doing.

This pr mostly works because the cuda compiler is doing pretty poorly here.