@drrros 2025-08-24
Master:
```
drros@tesla:~/llama.cpp$ ./build/bin/llama-bench --model /mnt/ds1nfs/codellamaweights/gpt-oss-20b-mxfp4.gguf -t 1 -fa 1 -b 4096 -ub 4096 -p 2048
ggml_cuda_init: GGML_CUDA_FORCE_MMQ:    no
ggml_cuda_init: GGML_CUDA_FORCE_CUBLAS: no
ggml_cuda_init: found 2 CUDA devices:
  Device 0: NVIDIA RTX A5000, compute capability 8.6, VMM: yes
  Device 1: NVIDIA RTX A5000, compute capability 8.6, VMM: yes
| model                          |       size |     params | backend    | ngl | threads | n_batch | n_ubatch | fa |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ------: | -------: | -: | --------------: | -------------------: |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |       1 |    4096 |     4096 |  1 |          pp2048 |       3830.22 ± 6.69 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |       1 |    4096 |     4096 |  1 |           tg128 |        138.25 ± 0.17 |

build: b730706a (6263)

``` 
This branch:
```
drros@tesla:~/llama.cpp$ ./build/bin/llama-bench --model /mnt/ds1nfs/codellamaweights/gpt-oss-20b-mxfp4.gguf -t 1 -fa 1 -b 4096 -ub 4096 -p 2048
ggml_cuda_init: GGML_CUDA_FORCE_MMQ:    no
ggml_cuda_init: GGML_CUDA_FORCE_CUBLAS: no
ggml_cuda_init: found 2 CUDA devices:
  Device 0: NVIDIA RTX A5000, compute capability 8.6, VMM: yes
  Device 1: NVIDIA RTX A5000, compute capability 8.6, VMM: yes
| model                          |       size |     params | backend    | ngl | threads | n_batch | n_ubatch | fa |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ------: | -------: | -: | --------------: | -------------------: |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |       1 |    4096 |     4096 |  1 |          pp2048 |       4083.86 ± 7.69 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |       1 |    4096 |     4096 |  1 |           tg128 |        158.43 ± 0.22 |

build: 7720366f (6212)

``` 
waiting for merge!