@mizuikk 2025-08-26
**Performance has improved a lot.**
<details>
<summary>Non-rigorous Performance Comparison for gpt-oss 20B MXFP4 MoE</summary>

ggml_cuda_init: GGML_CUDA_FORCE_MMQ:    no
ggml_cuda_init: GGML_CUDA_FORCE_CUBLAS: no
ggml_cuda_init: found 2 CUDA devices:
  Device 0: Tesla V100-SXM2-16GB, compute capability 7.0, VMM: yes
  Device 1: Tesla V100-SXM2-16GB, compute capability 7.0, VMM: yes
load_backend: loaded CUDA backend from /app/libggml-cuda.so
load_backend: loaded CPU backend from /app/libggml-cpu-haswell.so

## Batch Size Comparison

| n_batch | n_ubatch | Test    | t/s (Build e92734d5 (6250)) | t/s (Build 39842a7f (6281)) | Increment (%) |
| ------- | -------- | ------- | --------------------------- | --------------------------- | ------------- |
| 4096    | 2048     | pp2048  | 4033.44 ± 128.36            | 4080.68 ± 128.59            | 1.17          |
| 4096    | 2048     | pp8192  | 3572.80 ± 4.06              | 3608.89 ± 3.72              | 1.01          |
| 4096    | 2048     | pp16384 | 3043.20 ± 3.20              | 3078.45 ± 4.16              | 1.16          |
| 4096    | 2048     | pp32768 | 2352.16 ± 0.54              | 2382.57 ± 0.32              | 1.29          |
| 4096    | 2048     | tg128   | 125.71 ± 0.02               | 139.66 ± 0.02               | 11.10         |
| 4096    | 4096     | pp2048  | 4089.27 ± 11.14             | 4126.88 ± 10.73             | 0.92          |
| 4096    | 4096     | pp8192  | 3767.56 ± 10.57             | 3791.91 ± 10.15             | 0.65          |
| 4096    | 4096     | pp16384 | 3184.26 ± 2.14              | 3210.55 ± 1.21              | 0.83          |
| 4096    | 4096     | pp32768 | 2428.32 ± 1.06              | 2394.49 ± 107.44            | -1.39         |
| 4096    | 4096     | tg128   | 125.55 ± 0.02               | 139.57 ± 0.05               | 11.16         |

## Same talk Comparison

| t/s (Build e92734d5 (6250)) | t/s (Build 39842a7f (6281)) | Increment (%) |
| --------------------------- | --------------------------- | ------------- |
| 66.76                       | 71.27                       | 6.33          |
</details>